// samples/server/petstore/typescript-express-zod-server/builds/default/types.ts
/**
 * This code was generated by openapi-generator (typescript-express-zod-server).
 *
 * Changes to this file may cause incorrect behavior and will be lost if
 * the code is regenerated.
 */

/** Interface for the AuthorsService */
export interface AuthorsService {
  /** Create an author */
  createAuthor(params: CreateAuthorParams): Promise<Author>;

  /** Get an author by id */
  getAuthor(params: GetAuthorParams): Promise<Author>;

}

/** Interface for the CatalogService */
export interface CatalogService {
  /** Create a book */
  createBook(params: CreateBookParams): Promise<Book>;

  /** Delete a book */
  deleteBook(params: DeleteBookParams): Promise<void>;

  /** Get a book by id */
  getBook(params: GetBookParams): Promise<Book>;

  /** Get the latest event for a book (discriminated union response) */
  getLatestBookEvent(params: GetLatestBookEventParams): Promise<BookEvent>;

}

export type CreateAuthorParams = {
  /** author */
  author: Author;
};

export type CreateBookParams = {
  /** createBookRequest */
  createBookRequest: CreateBookRequest;
};

export type DeleteBookParams = {
  /** bookId */
  bookId: number;
};

export type GetAuthorParams = {
  /** authorId */
  authorId: number;
};

export type GetBookParams = {
  /** bookId */
  bookId: number;
  /** expand */
  expand?: boolean;
};

export type GetLatestBookEventParams = {
  /** bookId */
  bookId: number;
};

export type Author = {
  id?: number;
  name: string;
};

export type Genre = 'FICTION' | 'NONFICTION' | 'POETRY';

export type Book = {
  id?: number;
  title: string;
  genre?: Genre;
  author?: Author | null;
  pageCount?: number | null;
};

export type BookAddedEvent = {
  type: "BOOK_ADDED";
  bookId: number;
};

export type BookRemovedEvent = {
  type: "BOOK_REMOVED";
  bookId: number;
};

export type BookEvent = BookAddedEvent | BookRemovedEvent;

export function isBookAddedEvent(obj: BookEvent): obj is BookAddedEvent {
  return obj.type === "BOOK_ADDED";
}

export function isBookRemovedEvent(obj: BookEvent): obj is BookRemovedEvent {
  return obj.type === "BOOK_REMOVED";
}

export type CreateBookRequest = {
  title: string;
  genre?: Genre;
  authorId?: number;
};

export type ModelError = {
  message: string;
  code?: string;
};